# backend/docs/swagger.yaml
definitions:
...
  internal_files.StorageUsageResponse:
    properties:
      available_bytes:
        type: integer
      file_count:
        type: integer
      percent_used:
        type: number
      quota_bytes:
        type: integer
      quota_gb:
        type: number
      used_bytes:
        type: integer
      used_gb:
        type: number
    type: object
  internal_files.shareRequest:
    properties:
      access_mode:
        type: string
      file_id:
        type: string
      filename:
        type: string
      whitelist:
        items:
          type: string
        type: array
    type: object
...
  models.FileMetadata:
    properties:
      access_mode:
        type: string
      created_at:
        type: string
      encryption_type:
        type: string
      filename:
        type: string
      id:
        type: string
      owner_id:
        type: string
      public_cid:
        type: string
      size:
        type: integer
      storage_path:
        type: string
      updated_at:
        type: string
      whitelist:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
...
  /files:
    get:
      description: Lấy danh sách file của user hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FileMetadata'
            type: array
      security:
      - BearerAuth: []
      summary: List files
      tags:
      - files
  /files/{id}:
    delete:
      description: Xóa file trên R2 và metadata trong MongoDB
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete file
      tags:
      - files
  /files/{id}/download:
    get:
      description: Kiểm tra quyền và trả về presigned URL để download
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download file
      tags:
      - files
  /files/share:
    patch:
      consumes:
      - application/json
      description: Cập nhật access mode và/hoặc whitelist của file (chỉ owner)
      parameters:
      - description: Share settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_files.shareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FileMetadata'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Share file
      tags:
      - files
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload và lưu metadata file (stream mã hóa AES-GCM, không buffer
        toàn bộ vào RAM)
      parameters:
      - description: File cần upload
        in: formData
        name: file
        required: true
        type: file
      - description: Mật khẩu mã hóa file
        in: formData
        name: password
        required: true
        type: string
      - description: AES-128 | AES-192 | AES-256
        in: formData
        name: encryption_type
        type: string
      - description: public | private | whitelist
        in: formData
        name: access_mode
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FileMetadata'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload file
      tags:
      - files
  /storage/me:
    get:
      description: Lấy thông tin dung lượng đã dùng/quota của user hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_files.StorageUsageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my storage usage
      tags:
      - storage
swagger: "2.0"